Add more information to the generated GitHub build matrix
There are now two more fields available in each entry of the build
matrix: `oldest` and `newest`. Both fields are boolean values and
indicate whether the GHC version of the matrix entry is the newest
version or the oldest version respectively that was found in the
tested-with stanza of the Cabal file.